refactor!: established and followed the names of namespace, projects, assemblies, folders. BREAKING CHANGE
