+++ benchmarks/Ulid.Benchmarks/usings.cs
@@ -1,0 +1,3 @@
+global using System.Text;
+
+global using vm2.Providers;
