+++ src/Ulid/Ulid.csproj
@@ -2,6 +2,7 @@
 
     <PropertyGroup>
         <RootNamespace>vm2</RootNamespace>
+        <AssemblyName>vm2.Ulid</AssemblyName>
     </PropertyGroup>
 
     <PropertyGroup Label="NuGet">
