+++ src/Ulid/usings.cs
@@ -7,8 +7,10 @@ global using System.Diagnostics.CodeAnalysis;
 global using System.Numerics;
 global using System.Security.Cryptography;
 global using System.Text.Json;
+global using System.Text.Json.Serialization;
 
-global using vm2.UlidSerialization.SysJson;
+global using vm2.Serialization.SysJson;
+global using vm2.Providers;
 
 global using static System.Buffers.Binary.BinaryPrimitives;
 global using static vm2.Ulid;
