+++ test/Ulid.Tests/SysJson/UlidSysConverterTests.cs
@@ -1,9 +1,11 @@
 ﻿// SPDX-License-Identifier: MIT
 // Copyright (c) 2025-2026 Val Melamed
 
-namespace vm2.UlidType.Tests.SysJson;
+namespace vm2.Tests.Ulid.SysJson;
 
 using System.Text.Json;
+
+using vm2;
 
 [ExcludeFromCodeCoverage]
 public class UlidSysConverterTests
