+++ test/Ulid.Tests/Ulid.Tests.csproj
@@ -1,7 +1,8 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <RootNamespace>vm2.UlidType.Tests</RootNamespace>
+        <RootNamespace>vm2.Tests.Ulid</RootNamespace>
+        <AssemblyName>vm2.Ulid.Tests</AssemblyName>
     </PropertyGroup>
 
     <ItemGroup>
