+++ test/Ulid.Tests/UlidTests.cs
@@ -1,7 +1,9 @@
 ﻿// SPDX-License-Identifier: MIT
 // Copyright (c) 2025-2026 Val Melamed
 
-namespace vm2.UlidType.Tests;
+namespace vm2.Tests.Ulid;
+
+using vm2;
 
 [ExcludeFromCodeCoverage]
 public partial class UlidTests(ITestOutputHelper output) : TestBase(output)
