+++ test/UlidTool.Tests/UlidTool.Tests.csproj
@@ -1,7 +1,8 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <RootNamespace>vm2.UlidTool.Tests</RootNamespace>
+        <RootNamespace>vm2.Tests.UlidTool</RootNamespace>
+        <AssemblyName>vm2.UlidTool.Tests</AssemblyName>
     </PropertyGroup>
 
     <ItemGroup>
