+++ test/UlidTool.Tests/usings.cs
@@ -1,15 +1,6 @@
 ﻿// SPDX-License-Identifier: MIT
 // Copyright (c) 2025-2026 Val Melamed
 
-global using System.Diagnostics.CodeAnalysis;
-global using System.Runtime.Serialization;
-global using System.Text;
 global using System.Text.RegularExpressions;
 
 global using vm2.TestUtilities;
-
-global using vm2.Providers;
-
-global using static vm2.TestUtilities.TestUtilities;
-
-global using static vm2.Ulid;
